// Frontend/front-university/src/app/services/api/models/stundent-list-dto.ts
/* tslint:disable */
/* eslint-disable */
export interface StundentListDto {
  age: number;
  category:  string;
  city: string;
  comunity: string;
  country: string;
  courseName: string;
  dob: Date;
  firstName: string;
  email: string;
  id: number;
  lastName: string;
  state: string;
  street: string;
  zipCode: string;
}